TVG-97 Rename Reminder module
