+++ database/migration.py
@@ -6,7 +6,7 @@ from config import session
 from database.DatabaseService import DatabaseService
 from database.models.GuideModel import Guide
 from database.models.GuideEpisode import GuideEpisode
-from database.models.Reminder import Reminder
+from database.models.ReminderModel import Reminder
 from database.models.SearchItemModel import SearchItem
 from database.models.ShowDetailsModel import ShowDetails
 from database.models.ShowEpisodeModel import ShowEpisode
