+++ database/models/GuideModel.py
@@ -11,7 +11,7 @@ from aux_methods.types import ShowData
 from config import session
 from database import Base
 from database.models.GuideEpisode import GuideEpisode
-from database.models.Reminder import Reminder
+from database.models.ReminderModel import Reminder
 from database.models.SearchItemModel import SearchItem
 from database.models.ShowDetailsModel import ShowDetails
 from database.models.ShowEpisodeModel import ShowEpisode
