+++ tests/test_guide.py
@@ -443,7 +443,7 @@ class TestGuide(unittest.TestCase):
 
     @patch('sqlalchemy.orm.session.Session.commit')
     @patch('sqlalchemy.orm.session.Session.execute')
-    @patch('database.models.Reminder.get_reminder_by_show')
+    @patch('database.models.ReminderModel.Reminder.get_reminder_by_show')
     @patch('database.models.ShowEpisodeModel.ShowEpisode.search_for_episode')
     @patch('database.models.ShowDetailsModel.ShowDetails.get_show_by_title')
     @patch('database.models.SearchItemModel.SearchItem.get_active_searches')
